Create FaceDetection.java
FaceDection class converts the matrix image received from the controller class into a gray scale matrix image. It then runs a face-cascade classifier on the image to see if it can find a face and if a face is found, calculate the center of gravity (COG) of the detected face and return those two calculated points to the controller class.