Create MotionDetection.java
MotionDetection class converts the received matrix image into a gray scale matrix image and then compares it to the last acquired image, which was saved to memory, to calculate if motion is detected between the images. This is called image differentiating of consecutive video frames. If motion is detected between the two images, the center of gravity of the detected motion is calculated and the class returns these two points, otherwise a null is returned.